--- src/write-the-docs/Tests/Utilities/__init__.py
@@ -1,2 +1,0 @@
-from write_the_docs.Utilities.command_line import shell, set_permissions
-from write_the_docs.Utilities.logging import configure_logging
