+++ src/writethedocs/Classes/black.py
@@ -1,19 +1,19 @@
 import logging
 import traceback
 from pathlib import Path
-from write_the_docs.Utilities import shell
+from ..Utilities import shell
 
 
-class Flake8:
+class Black:
     """
-    Class to configure and run flake8.
+    Class to configure and run black formatter.
 
     Parameters
     ----------
     source_code_path: str | Path
         Source code root directory.
-    entry_file: str | Path
-        Line length to pass to flake8 as max-line-length.
+    line_length: int
+        Line length to pass to black formatter as line-length.
     additional_params: list[str]
         Optional additional paramters for Sphinx autodoc.
 
@@ -22,7 +22,7 @@ class Flake8:
     source_code_path: str | Path
         Source code root directory.
     line_length: int
-        Line length to pass to flake8 as max-line-length.
+        Line length to pass to black formatter as line-length.
     logger: Logger
         Logger to pass messages to.
     additional_params: list[str]
@@ -40,30 +40,28 @@ class Flake8:
             self.source_code_path = source_code_path
             self.line_length = line_length
             self.additional_params = additional_params
-
         except Exception as e:
-            msg = f"Error in flake8 initialisation: {traceback.format_exc()}"
+            msg = f"Error in black initialisation: {traceback.format_exc()}"
             self.logger.error(msg)
             raise e
 
     def run(self) -> None:
-        """Run flake8 on PowerShell."""
+        """Run black on PowerShell."""
         try:
             out, err, exit_code = shell(
                 [
                     "powershell.exe",
-                    "flake8",
-                    f"--max-line-length={self.line_length}",
-                    "--per-file-ignores='__init__.py:F401'",
+                    "black",
                     f"{self.source_code_path}",
-                    *self.additional_params,
+                    "--line-length",
+                    f"{self.line_length}",
                 ],
                 self.logger,
                 False,
+                *self.additional_params,
             )
             return out, err, exit_code
-
         except Exception as e:
-            msg = f"Error in flake8 run: {traceback.format_exc()}"
+            msg = f"Error in black run: {traceback.format_exc()}"
             self.logger.error(msg)
             raise e
