+++ src/writethedocs/Classes/flake8.py
@@ -1,19 +1,19 @@
 import logging
 import traceback
 from pathlib import Path
-from write_the_docs.Utilities import shell
+from ..Utilities import shell
 
 
-class Black:
+class Flake8:
     """
-    Class to configure and run black formatter.
+    Class to configure and run flake8.
 
     Parameters
     ----------
     source_code_path: str | Path
         Source code root directory.
-    line_length: int
-        Line length to pass to black formatter as line-length.
+    entry_file: str | Path
+        Line length to pass to flake8 as max-line-length.
     additional_params: list[str]
         Optional additional paramters for Sphinx autodoc.
 
@@ -22,7 +22,7 @@ class Black:
     source_code_path: str | Path
         Source code root directory.
     line_length: int
-        Line length to pass to black formatter as line-length.
+        Line length to pass to flake8 as max-line-length.
     logger: Logger
         Logger to pass messages to.
     additional_params: list[str]
@@ -40,28 +40,30 @@ class Black:
             self.source_code_path = source_code_path
             self.line_length = line_length
             self.additional_params = additional_params
+
         except Exception as e:
-            msg = f"Error in black initialisation: {traceback.format_exc()}"
+            msg = f"Error in flake8 initialisation: {traceback.format_exc()}"
             self.logger.error(msg)
             raise e
 
     def run(self) -> None:
-        """Run black on PowerShell."""
+        """Run flake8 on PowerShell."""
         try:
             out, err, exit_code = shell(
                 [
                     "powershell.exe",
-                    "black",
+                    "flake8",
+                    f"--max-line-length={self.line_length}",
+                    "--per-file-ignores='__init__.py:F401'",
                     f"{self.source_code_path}",
-                    "--line-length",
-                    f"{self.line_length}",
+                    *self.additional_params,
                 ],
                 self.logger,
                 False,
-                *self.additional_params,
             )
             return out, err, exit_code
+
         except Exception as e:
-            msg = f"Error in black run: {traceback.format_exc()}"
+            msg = f"Error in flake8 run: {traceback.format_exc()}"
             self.logger.error(msg)
             raise e
