+++ src/writethedocs/Classes/unit_tests.py
@@ -1,7 +1,7 @@
 import logging
 import traceback
 from pathlib import Path
-from write_the_docs.Utilities import shell, configure_logging
+from ..Utilities import shell
 
 
 class UnitTests:
