+++ src/writethedocs/Tests/Utilities/__init__.py
@@ -1,0 +1,2 @@
+from ...Utilities.command_line import shell, set_permissions
+from ...Utilities.logging import configure_logging
