+++ src/writethedocs/Tests/Utilities/test_logging.py
@@ -1,9 +1,7 @@
 import unittest
-import shutil
-import coloredlogs
 import logging
 from pathlib import Path
-from write_the_docs.Utilities import str_to_html, configure_logging
+from ...Utilities import str_to_html, configure_logging
 
 temp_path = Path(Path(__file__).parent.parent, "temp", "logs")
 temp_path.mkdir(exist_ok=True, parents=True)
